Extract remove-funciton module from MatmulEpilogueFusion
